Additional optimizations
