+++ tensorqtl/cis.py
@@ -53,6 +53,19 @@ def calculate_cis_nominal(genotypes_t, phenotype_t, residualizer):
     return tstat_t, slope_t, slope_se_t, maf_t, ma_samples_t, ma_count_t
 
 
+def calculate_cis_permutations(genotypes_t, phenotype_t, residualizer, permutation_ix_t):
+    """Calculate nominal and empirical correlations"""
+    permutations_t = phenotype_t[permutation_ix_t]
+
+    r_nominal_t, std_ratio_t = calculate_corr(genotypes_t, phenotype_t.reshape(1,-1), residualizer, return_sd=True)
+    r_nominal_t = r_nominal_t.squeeze()
+    corr_t = calculate_corr(genotypes_t, permutations_t, residualizer).pow(2)  # genotypes x permutations
+    r2_perm_t,_ = corr_t[~torch.isnan(corr_t).any(1),:].max(0)
+
+    ix = r_nominal_t.pow(2).argmax()
+    return r_nominal_t[ix], std_ratio_t[ix], ix, r2_perm_t, genotypes_t[ix]  # , r_nominal_t.shape[0]
+
+
 def map_nominal(genotype_df, variant_df, phenotype_df, phenotype_pos_df, covariates_df, prefix,
                 window=1000000, group_s=None, output_dir='.', logger=None):
     """
@@ -185,7 +198,7 @@ def map_nominal_interaction(genotype_df, variant_df, phenotype_df, phenotype_pos
         chr_res_df = []
         for k, (phenotype, genotypes, genotype_range, phenotype_id) in enumerate(igc.generate_data(chrom=chrom), k+1):
             # copy genotypes to GPU
-            phenotype_t = torch.tensor(phenotype, dtype=torch.float).to(device).unsqueeze(0)
+            phenotype_t = torch.tensor(phenotype, dtype=torch.float).to(device)
             genotypes_t = torch.tensor(genotypes, dtype=torch.float).to(device)
             genotypes_t = genotypes_t[:,genotype_ix_t]
             impute_mean(genotypes_t)
@@ -240,7 +253,7 @@ def map_nominal_interaction(genotype_df, variant_df, phenotype_df, phenotype_pos
             chr_res_df['pval_i'] = 2*stats.t.cdf(-chr_res_df['pval_i'].abs(), dof)
             chr_res_df['pval_gi'] = 2*stats.t.cdf(-chr_res_df['pval_gi'].abs(), dof)
             print('  * writing output')
-            chr_res_df.to_parquet(os.path.join(output_dir, '{}.cis_qtl_pairsX.{}.parquet'.format(prefix, chrom)))
+            chr_res_df.to_parquet(os.path.join(output_dir, '{}.cis_qtl_pairs.{}.parquet'.format(prefix, chrom)))
 
     best_assoc = pd.concat(best_assoc, axis=1).T.set_index('phenotype_id').infer_objects()
     best_assoc['pval_g'] = 2*stats.t.cdf(-best_assoc['pval_g'].abs(), dof)
@@ -251,4 +264,302 @@ def map_nominal_interaction(genotype_df, variant_df, phenotype_df, phenotype_pos
     logger.write('done.')
 
 
+def prepare_cis_output(r_nominal, r2_perm, std_ratio, g, num_var, dof, variant_id, tss_distance, phenotype_id, nperm=10000):
+    """Return nominal p-value, allele frequencies, etc. as pd.Series"""
+    r2_nominal = r_nominal*r_nominal
+    pval_perm = (np.sum(r2_perm>=r2_nominal)+1) / (nperm+1)
 
+    slope = r_nominal * std_ratio
+    tstat2 = dof * r2_nominal / (1 - r2_nominal)
+    slope_se = np.abs(slope) / np.sqrt(tstat2)
+
+    n2 = 2*len(g)
+    maf = np.sum(g) / n2
+    if maf <= 0.5:
+        ref_factor = 1
+        ma_samples = np.sum(g>0.5)
+        ma_count = np.sum(g[g>0.5])
+    else:
+        maf = 1-maf
+        ref_factor = -1
+        ma_samples = np.sum(g<1.5)
+        ma_count = n2 - np.sum(g[g>0.5])
+
+    res_s = pd.Series(OrderedDict([
+        ('num_var', num_var),
+        ('beta_shape1', np.NaN),
+        ('beta_shape2', np.NaN),
+        ('true_df', np.NaN),
+        ('pval_true_df', np.NaN),
+        ('variant_id', variant_id),
+        ('tss_distance', tss_distance),
+        ('ma_samples', ma_samples),
+        ('ma_count', ma_count),
+        ('maf', maf),
+        ('ref_factor', ref_factor),
+        ('pval_nominal', pval_from_corr(r2_nominal, dof)),
+        ('slope', slope),
+        ('slope_se', slope_se),
+        ('pval_perm', pval_perm),
+        ('pval_beta', np.NaN),
+    ]), name=phenotype_id)
+    return res_s
+
+
+def _process_group_permutations(buf):
+    """
+    Merge results for grouped phenotypes
+
+    buf: [r_nominal, std_ratio, var_ix, r2_perm, g, num_var, phenotype_id]
+    """
+    # select phenotype with strongest nominal association
+    max_ix = np.argmax(np.abs([b[0] for b in buf]))
+    r_nominal, std_ratio, var_ix = buf[max_ix][:3]
+    g, num_var, phenotype_id = buf[max_ix][4:]
+    # select best phenotype correlation for each permutation
+    r2_perm = np.max([b[3] for b in buf], 0)
+    return r_nominal, std_ratio, var_ix, r2_perm, g, num_var, phenotype_id
+
+
+def map_cis(genotype_df, variant_df, phenotype_df, phenotype_pos_df, covariates_df,
+            group_s=None, beta_approx=True, nperm=10000, window=1000000, logger=None, seed=None):
+    """Run cis-QTL mapping"""
+
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+
+    assert np.all(phenotype_df.columns==covariates_df.index)
+    if logger is None:
+        logger = SimpleLogger()
+
+    logger.write('cis-QTL mapping: empirical p-values for phenotypes')
+    logger.write('  * {} samples'.format(phenotype_df.shape[1]))
+    logger.write('  * {} phenotypes'.format(phenotype_df.shape[0]))
+    if group_s is not None:
+        logger.write('  * {} phenotype groups'.format(len(group_s.unique())))
+        group_dict = group_s.to_dict()
+    logger.write('  * {} covariates'.format(covariates_df.shape[1]))
+    logger.write('  * {} variants'.format(genotype_df.shape[0]))
+
+    covariates_t = torch.tensor(covariates_df.values, dtype=torch.float32).to(device)
+    residualizer = Residualizer(covariates_t)
+    del covariates_t
+
+    genotype_ix = np.array([genotype_df.columns.tolist().index(i) for i in phenotype_df.columns])
+    genotype_ix_t = torch.from_numpy(genotype_ix).to(device)
+    dof = phenotype_df.shape[1] - 2 - covariates_df.shape[1]
+
+    # permutation indices
+    n_samples = phenotype_df.shape[1]
+    ix = np.arange(n_samples)
+    if seed is not None:
+        np.random.seed(seed)
+    permutation_ix_t = torch.LongTensor(np.array([np.random.permutation(ix) for i in range(nperm)])).to(device)
+
+    res_df = []
+    igc = genotypeio.InputGeneratorCis(genotype_df, variant_df, phenotype_df, phenotype_pos_df, window=window)
+    start_time = time.time()
+    if group_s is None:
+        for k, (phenotype, genotypes, genotype_range, phenotype_id) in enumerate(igc.generate_data(), 1):
+            # copy genotypes to GPU
+            phenotype_t = torch.tensor(phenotype, dtype=torch.float).to(device)
+            genotypes_t = torch.tensor(genotypes, dtype=torch.float).to(device)
+            genotypes_t = genotypes_t[:,genotype_ix_t]
+            impute_mean(genotypes_t)
+
+            res = calculate_cis_permutations(genotypes_t, phenotype_t, residualizer, permutation_ix_t)
+            r_nominal, std_ratio, var_ix, r2_perm, g = [i.cpu().numpy() for i in res]
+            var_ix = genotype_range[var_ix]
+            variant_id = variant_df.index[var_ix]
+            tss_distance = variant_df['pos'].values[var_ix] - igc.phenotype_tss[phenotype_id]
+            res_s = prepare_cis_output(r_nominal, r2_perm, std_ratio, g, genotypes.shape[0], dof, variant_id, tss_distance, phenotype_id, nperm=nperm)
+            if beta_approx:
+                res_s[['pval_beta', 'beta_shape1', 'beta_shape2', 'true_df', 'pval_true_df']] = calculate_beta_approx_pval(r2_perm, r_nominal*r_nominal, dof)
+            res_df.append(res_s)
+            print('\r  * computing permutations for phenotype {}/{}'.format(k, igc.n_phenotypes), end='')
+        print()
+    else:
+        n_groups = len(igc.phenotype_df.index.map(group_dict).unique())
+        buf = []
+        processed_groups = 0
+        previous_group = None
+        for k, (phenotype, genotypes, genotype_range, phenotype_id) in enumerate(igc.generate_data(), 1):
+            # copy genotypes to GPU
+            phenotype_t = torch.tensor(phenotype, dtype=torch.float).to(device)
+            genotypes_t = torch.tensor(genotypes, dtype=torch.float).to(device)
+            genotypes_t = genotypes_t[:,genotype_ix_t]
+            impute_mean(genotypes_t)
+
+            group_id = group_dict[phenotype_id]
+            res = calculate_cis_permutations(genotypes_t, phenotype_t, residualizer, permutation_ix_t)
+            res = [i.cpu().numpy() for i in res]  # r_nominal, std_ratio, var_ix, r2_perm, g
+            res[2] = genotype_range[var_ix]
+            if (group_id != previous_group and len(buf)>0):  # new group, process previous
+                r_nominal, std_ratio, var_ix, r2_perm, g, num_var, phenotype_id = _process_group_permutations(buf)
+                variant_id = variant_df.index[var_ix]
+                tss_distance = variant_df['pos'].values[var_ix] - igc.phenotype_tss[phenotype_id]
+                res_s = prepare_cis_output(r_nominal, r2_perm, std_ratio, g, num_var, dof, variant_id, tss_distance, phenotype_id, nperm=nperm)
+                res_s[['pval_beta', 'beta_shape1', 'beta_shape2', 'true_df', 'pval_true_df']] = calculate_beta_approx_pval(r2_perm, r_nominal*r_nominal, dof)
+                res_s['group_id'] = group_id
+                res_s['group_size'] = len(buf)
+                res_df.append(res_s)
+                processed_groups += 1
+                print('\r  * computing permutations for phenotype group {}/{}'.format(processed_groups, n_groups), end='')
+                # reset
+                buf = [ires]
+            else:
+                buf.append(res + [genotypes.shape[0], phenotype_id])
+            previous_group = group_id
+        # process last group
+        r_nominal, std_ratio, var_ix, r2_perm, g, num_var, phenotype_id = _process_group_permutations(buf)
+        variant_id = variant_df.index[var_ix]
+        tss_distance = variant_df['pos'].values[var_ix] - igc.phenotype_tss[phenotype_id]
+        res_s = prepare_cis_output(r_nominal, r2_perm, std_ratio, g, num_var, dof, variant_id, tss_distance, phenotype_id, nperm=nperm)
+        res_s[['pval_beta', 'beta_shape1', 'beta_shape2', 'true_df', 'pval_true_df']] = calculate_beta_approx_pval(r2_perm, r_nominal*r_nominal, dof)
+        res_s['group_id'] = group_id
+        res_s['group_size'] = len(buf)
+        res_df.append(res_s)
+        processed_groups += 1
+        print('\r  * computing permutations for phenotype group {}/{}'.format(processed_groups, n_groups), end='\n')
+
+    res_df = pd.concat(res_df, axis=1).T
+    res_df.index.name = 'phenotype_id'
+    logger.write('  Time elapsed: {:.2f} min'.format((time.time()-start_time)/60))
+    logger.write('done.')
+    return res_df.astype(output_dtype_dict).infer_objects()
+
+
+def map_independent(genotype_df, variant_df, cis_df, phenotype_df, phenotype_pos_df, covariates_df,
+                    group_s=None, fdr=0.05, fdr_col='qval', nperm=10000, window=1000000, logger=None, seed=None):
+    """
+    Run independent cis-QTL mapping (forward-backward regression)
+
+    cis_df: output from map_cis, annotated with q-values (calculate_qvalues)
+    """
+
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+
+    assert np.all(phenotype_df.index==phenotype_pos_df.index)
+    if logger is None:
+        logger = SimpleLogger()
+
+    signif_df = cis_df[cis_df[fdr_col]<=fdr].copy()
+    signif_df = signif_df[[
+        'num_var', 'beta_shape1', 'beta_shape2', 'true_df', 'pval_true_df',
+        'variant_id', 'tss_distance', 'ma_samples', 'ma_count', 'maf', 'ref_factor',
+        'pval_nominal', 'slope', 'slope_se', 'pval_perm', 'pval_beta',
+    ]]
+    signif_threshold = signif_df['pval_beta'].max()
+    # subset significant phenotypes
+    ix = signif_df.index[signif_df.index.isin(phenotype_df.index)]
+    phenotype_df = phenotype_df.loc[ix]
+    phenotype_pos_df = phenotype_pos_df.loc[ix]
+
+    logger.write('cis-QTL mapping: conditionally independent variants')
+    logger.write('  * {} samples'.format(phenotype_df.shape[1]))
+    logger.write('  * {} significant phenotypes'.format(signif_df.shape[0]))
+    if group_s is not None:
+        # logger.write('  * {} phenotype groups'.format(len(group_s.unique())))
+        # group_dict = group_s.to_dict()
+        raise ValueError('Grouped mode not yet implemented.')
+    logger.write('  * {} covariates'.format(covariates_df.shape[1]))
+    logger.write('  * {} variants'.format(genotype_df.shape[0]))
+
+    # print('Significance threshold: {}'.format(signif_threshold))
+
+    # covariates_t = torch.tensor(covariates_df.values, dtype=torch.float32).to(device)
+    # residualizer = Residualizer(covariates_t)
+    # del covariates_t
+
+    genotype_ix = np.array([genotype_df.columns.tolist().index(i) for i in phenotype_df.columns])
+    genotype_ix_t = torch.from_numpy(genotype_ix).to(device)
+    dof = phenotype_df.shape[1] - 2 - covariates_df.shape[1]
+    ix_dict = {i:k for k,i in enumerate(genotype_df.index)}
+
+    # permutation indices
+    n_samples = phenotype_df.shape[1]
+    ix = np.arange(n_samples)
+    if seed is not None:
+        np.random.seed(seed)
+    permutation_ix_t = torch.LongTensor(np.array([np.random.permutation(ix) for i in range(nperm)])).to(device)
+
+    res_df = []
+    igc = genotypeio.InputGeneratorCis(genotype_df, variant_df, phenotype_df, phenotype_pos_df, window=window)
+    start_time = time.time()
+    if group_s is None:
+        for k, (phenotype, genotypes, genotype_range, phenotype_id) in enumerate(igc.generate_data(), 1):
+            # copy genotypes to GPU
+            phenotype_t = torch.tensor(phenotype, dtype=torch.float).to(device).unsqueeze(0)
+            genotypes_t = torch.tensor(genotypes, dtype=torch.float).to(device)
+            genotypes_t = genotypes_t[:,genotype_ix_t]
+            impute_mean(genotypes_t)
+
+            # 1) forward pass
+            forward_df = [signif_df.loc[phenotype_id]]  # initialize results with top variant
+            covariates = covariates_df.values.copy()  # initialize covariates
+            dosage_dict = {}
+            while True:
+                # add variant to covariates
+                variant_id = forward_df[-1]['variant_id']
+                ig = genotype_df.values[ix_dict[variant_id], genotype_ix]
+                dosage_dict[variant_id] = ig
+                covariates = np.hstack([covariates, ig.reshape(-1,1)]).astype(np.float32)
+                dof = phenotype_df.shape[1] - 2 - covariates.shape[1]
+                covariates_t = torch.tensor(covariates, dtype=torch.float32).to(device)
+                residualizer = Residualizer(covariates_t)
+                del covariates_t
+
+                res = calculate_cis_permutations(genotypes_t, phenotype_t, residualizer, permutation_ix_t)
+                r_nominal, std_ratio, var_ix, r2_perm, g = [i.cpu().numpy() for i in res]
+
+                x = calculate_beta_approx_pval(r2_perm, r_nominal*r_nominal, dof)
+                # add to list if significant
+                if x[0] <= signif_threshold:
+                    var_ix = genotype_range[var_ix]
+                    variant_id = variant_df.index[var_ix]
+                    tss_distance = variant_df['pos'].values[var_ix] - igc.phenotype_tss[phenotype_id]
+                    res_s = prepare_cis_output(r_nominal, r2_perm, std_ratio, g, genotypes.shape[0], dof, variant_id, tss_distance, phenotype_id, nperm=nperm)
+                    res_s[['pval_beta', 'beta_shape1', 'beta_shape2', 'true_df', 'pval_true_df']] = x
+                    forward_df.append(res_s)
+                else:
+                    break
+            forward_df = pd.concat(forward_df, axis=1).T
+            dosage_df = pd.DataFrame(dosage_dict)
+
+            # 2) backward pass
+            if forward_df.shape[0]>1:
+                back_df = []
+                variant_set = set()
+                for k,i in enumerate(forward_df['variant_id'], 1):
+                    covariates = np.hstack([
+                        covariates_df.values,
+                        dosage_df[np.setdiff1d(forward_df['variant_id'], i)].values,
+                    ])
+                    dof = phenotype_df.shape[1] - 2 - covariates.shape[1]
+                    covariates_t = torch.tensor(covariates, dtype=torch.float32).to(device)
+                    residualizer = Residualizer(covariates_t)
+                    del covariates_t
+
+                    res = calculate_cis_permutations(genotypes_t, phenotype_t, residualizer, permutation_ix_t)
+                    r_nominal, std_ratio, var_ix, r2_perm, g = [i.cpu().numpy() for i in res]
+                    var_ix = genotype_range[var_ix]
+                    variant_id = variant_df.index[var_ix]
+                    x = calculate_beta_approx_pval(r2_perm, r_nominal*r_nominal, dof)
+                    if x[0] <= signif_threshold and variant_id not in variant_set:
+                        tss_distance = variant_df['pos'].values[var_ix] - igc.phenotype_tss[phenotype_id]
+                        res_s = prepare_cis_output(r_nominal, r2_perm, std_ratio, g, num_var, dof, variant_id, tss_distance, phenotype_id, nperm=nperm)
+                        res_s[['pval_beta', 'beta_shape1', 'beta_shape2', 'true_df', 'pval_true_df']] = x
+                        res_s['rank'] = k
+                        back_df.append(res_s)
+                        variant_set.add(variant_id)
+                if len(back_df)>0:
+                    res_df.append(pd.concat(back_df, axis=1).T)
+            else:  # single independent variant
+                forward_df['rank'] = 1
+                res_df.append(forward_df)
+            print('\r  * computing independent QTL for phenotype {}/{}'.format(j, igc.n_phenotypes), end='')
+        print()
+    res_df = pd.concat(res_df, axis=0)
+    res_df.index.name = 'phenotype_id'
+    logger.write('  Time elapsed: {:.2f} min'.format((time.time()-start_time)/60))
+    logger.write('done.')
+    return res_df.reset_index().astype(output_dtype_dict)
